+++ include/Particle.hpp
@@ -1,0 +1,12 @@
+#pragma once
+
+class Particle {
+ public:
+  Particle();
+
+  double originX;
+  double originY;
+  double velocityX;
+  double velocityY;
+  int speed;
+};