+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.1] - 2025-05-11
+
+### Added
+
+-   More "magic numbers" as constants
+
+### Changed
+
+-   Switched `Explosion` and `Smoke` particles from dynamic C-style arrays to `std::array` of `Particle`
+
+### Fixed
+
+-   `UveDX::showError` string
+
 ## [0.3.0] - 2025-05-11
 
 ### Added